Fix pre-existing ruff lint and formatting issues
Apply ruff check --fix and ruff format across the codebase.
Remaining manual fixes: PLW1510 (check=False), SIM117 (combine with),
SIM102 (combine if), S110 (noqa), SIM115 (context manager).
All 187 tests pass.

+++ tools/usage.py
@@ -7,7 +7,6 @@ import math
 import threading
 from typing import Any
 
-
 _EXCLUDED_TOOLS = {"wiki_usage"}
 _USAGE_LOCK = threading.Lock()
 _SESSION_INPUT_TOKENS = 0
@@ -25,7 +24,7 @@ def _estimate_tokens(payload: Any) -> int:
         text = str(payload)
     if not text:
         return 0
-    return int(math.ceil(len(text) / 4.0))
+    return math.ceil(len(text) / 4.0)
 
 
 def record_tool_input(tool_name: str, payload: Any) -> int:
